requirements.txt
flask
flask_cors
flask_socketio
mysql-connector-python
google-generativeai
python-dotenv
uuid